+++ server/model/IPhpmdSettingsModel.ts
@@ -6,6 +6,13 @@
  */
 interface IPhpmdSettingsModel {
     /**
+     * Enabled
+     *
+     * @property {boolean}
+     */
+    enabled: boolean;
+    
+    /**
      * PHP mess detector command
      *
      * @property {string}
@@ -22,9 +29,16 @@ interface IPhpmdSettingsModel {
     /**
      * Verbose mode
      *
-     * @property {string}
+     * @property {boolean}
      */
     verbose: boolean;
+
+    /**
+     * Clear errors on close
+     * 
+     * @property {boolean}
+     */
+    clearOnClose: boolean;
 }
 
 export default IPhpmdSettingsModel;